backend/requirements.txt
flask
firebase-admin
pandas
numpy
scikit-learn
flask-cors
matplotlib
seaborn
openpyxl
Pillow
